refactor: ➖ Remove useless setup [skip ci]
I removed the setup.py as it is no longer needed and messed up my tests.
If you have trouble loading the python files, just add in a .env with the path going to the txt file you downloaded from advent or just replace the path that loads the environment variable.

Signed-off-by: Kami <harshimymommy@gmail.com>